Consolidated option methods into trait for reusability
